requirements.txt
a2a-sdk
click>=8.1.8
dotenv>=0.9.9
httpx>=0.28.1
google-genai>=1.9.0
google-adk>=0.0.3
pydantic>=2.11.4
python-dotenv>=1.1.0
uvicorn>=0.30.0

a2a-sdk
click>=8.1.8
python-dotenv>=1.1.0
uvicorn>=0.30.0
langgraph>=0.2.0
langchain-openai>=0.2.0
langchain-core>=0.3.0
requests>=2.31.0

a2a-sdk
click>=8.1.8
python-dotenv>=1.1.0
uvicorn>=0.30.0
langgraph>=0.2.0
langchain-openai>=0.2.0
langchain-core>=0.3.0
requests>=2.31.0